fix gem and bundler configs
After thinking about the BUNDLE_PATH thing for a
bit, I'm not really sure what it was helping with.
Yeah, the gems were local, but so what? I still
had to run `bundle exec` for everything, and
that's a pseudo-problem solved easily enough with
aliases and wrappers.

The final motivation for this is giving RubyMine a
shot, and it's got a bug that doesn't understand
per-project gem directories managed by bundler:

http://youtrack.jetbrains.com/issue/RUBY-14542

So, all the gems are in one place now, which is
turning out not to be that big of a deal.